+++ src/agilab/apps/builtin/data_quality_gate_project/src/data_quality_gate/compat/__init__.py
@@ -1,0 +1,1 @@
+"""Compatibility helpers for classified data quality gate modules."""
