+++ src/agilab/apps/builtin/data_quality_gate_project/src/data_quality_gate/data_quality_gate.py
@@ -1,111 +1,9 @@
-"""Manager for the built-in data quality gate app."""
+"""Compatibility shim for the classified data quality gate package layout."""
 
 from __future__ import annotations
 
-import logging
-import shutil
-from pathlib import Path
-from typing import Any
+from data_quality_gate.compat.module_shim import activate_compat_module
 
-from pydantic import ValidationError
+_TARGET_MODULE = "data_quality_gate.runtime.data_quality_gate"
 
-from agi_node.agi_dispatcher import BaseWorker, WorkDispatcher
-
-from .app_args import (
-    ArgsOverrides,
-    DataQualityGateArgs,
-    dump_args,
-    ensure_defaults,
-    filter_arg_overrides,
-    load_args,
-    merge_args,
-    safe_reset_path,
-    share_root_from_env,
-)
-
-logger = logging.getLogger(__name__)
-
-
-class DataQualityGate(BaseWorker):
-    """Manager that dispatches one deterministic data-quality gate run."""
-
-    worker_vars: dict[str, Any] = {}
-
-    def __init__(
-        self,
-        env,
-        args: DataQualityGateArgs | None = None,
-        **kwargs: ArgsOverrides,
-    ) -> None:
-        self.env = env
-        self._ensure_managed_pc_share_dir(env)
-        self.verbose = int(kwargs.pop("verbose", getattr(env, "verbose", 0) or 0))
-        arg_overrides = filter_arg_overrides(kwargs)
-
-        if args is None:
-            try:
-                args = DataQualityGateArgs(**arg_overrides)
-            except ValidationError as exc:
-                raise ValueError(f"Invalid DataQualityGate arguments: {exc}") from exc
-        elif arg_overrides:
-            args = merge_args(args, arg_overrides)
-
-        self.args = ensure_defaults(args, env=env)
-        self.args = self._apply_managed_pc_paths(self.args)
-        self.data_out = env.resolve_share_path(self.args.data_out)
-
-        if self.args.reset_target and self.data_out.exists():
-            reset_path = safe_reset_path(self.data_out, share_root=share_root_from_env(env), label="data_out")
-            shutil.rmtree(reset_path, ignore_errors=True, onerror=WorkDispatcher._onerror)
-        self.data_out.mkdir(parents=True, exist_ok=True)
-        self.analysis_artifact_dir.mkdir(parents=True, exist_ok=True)
-
-    @property
-    def analysis_artifact_dir(self) -> Path:
-        export_root = Path(getattr(self.env, "AGILAB_EXPORT_ABS", Path.home() / "export"))
-        target = str(
-            getattr(self.env, "target", "")
-            or getattr(self.env, "app", "")
-            or getattr(self.env, "active_app", "")
-            or "data_quality_gate_project"
-        )
-        return export_root / target / "data_quality_gate"
-
-    @classmethod
-    def from_toml(
-        cls,
-        env,
-        settings_path: str | Path = "app_settings.toml",
-        section: str = "args",
-        **overrides: ArgsOverrides,
-    ) -> "DataQualityGate":
-        base = load_args(settings_path, section=section)
-        merged = ensure_defaults(merge_args(base, overrides or None), env=env)
-        return cls(env, args=merged)
-
-    def to_toml(
-        self,
-        settings_path: str | Path = "app_settings.toml",
-        section: str = "args",
-        create_missing: bool = True,
-    ) -> None:
-        dump_args(self.args, settings_path, section=section, create_missing=create_missing)
-
-    def as_dict(self) -> dict[str, Any]:
-        return self.args.model_dump(mode="json")
-
-    def build_distribution(self, workers):
-        try:
-            worker_count = max(1, int(workers))
-        except (TypeError, ValueError):
-            worker_count = 1
-        work_plan = [[["data_quality_gate"]]] + [[] for _ in range(worker_count - 1)]
-        metadata = [[{"run": "data_quality_gate", "work_items": 1}]] + [[] for _ in range(worker_count - 1)]
-        return work_plan, metadata, "run", "work_items", "items"
-
-
-class DataQualityGateApp(DataQualityGate):
-    """Compatibility class with the descriptive *App suffix."""
-
-
-__all__ = ["DataQualityGate", "DataQualityGateApp"]
+activate_compat_module(__name__, _TARGET_MODULE)
