+++ src/agilab/apps/builtin/data_quality_gate_project/src/data_quality_gate/domain/__init__.py
@@ -1,0 +1,1 @@
+"""Domain helpers for the built-in data quality gate app."""
